refactor: move MuiDatePickerExample to examples/V7 and update imports
