+++ app/tsconfig.json
@@ -16,5 +16,5 @@
     "noEmit": true,
     "jsx": "react"
   },
-  "include": ["./src"]
+  "include": ["./src", "../examples/V7/mui-date-picker.tsx"]
 }
